update monitoriing intervals to float

+++ ada.py
@@ -107,7 +107,7 @@ def main():
     mode_monitor_shared_var = SharedVariable() # create thread safe variable for ADA to track
     mode_monitor_config = {
     "name": os.getenv("MODE_MONITOR_NAME", "default_ModeMonitor"),
-    "monitoring_interval": int(os.getenv("MODE_MONITOR_INTERVAL", 1)),
+    "monitoring_interval": float(os.getenv("MODE_MONITOR_INTERVAL", 1)),
     "threading_shared_var": mode_monitor_shared_var,
     "switch_reader": mode_switch
     }
@@ -133,7 +133,7 @@ def main():
     door_monitor_shared_var = SharedVariable() # create thread safe variable for ADA to track
     door_monitor_config = {
     "name": os.getenv("DOOR_MONITOR_NAME", "default_DoorMonitor"),
-    "monitoring_interval": int(os.getenv("DOOR_MONITOR_INTERVAL", 30)),
+    "monitoring_interval": float(os.getenv("DOOR_MONITOR_INTERVAL", 30)),
     "threading_shared_var": door_monitor_shared_var,
     "switch_reader": door_reed_switch
     }
@@ -149,7 +149,7 @@ def main():
     rfid_monitor_shared_var = SharedVariable() # create thread safe variable for ADA to track
     rfid_monitor_config = {
     "name": os.getenv("RFID_MONITOR_NAME", "default_RfidMonitor"),
-    "monitoring_interval": int(os.getenv("RFID_MONITOR_INTERVAL", 5)),
+    "monitoring_interval": float(os.getenv("RFID_MONITOR_INTERVAL", 5)),
     "threading_shared_var": rfid_monitor_shared_var,
     "mfrc522_reader": rfid_reader
     }
